folder_branch_ops: extra debugging for an op that's missing a path
I wasn't able to reproduce the panic described in HOTPOT-803 in a
test, so this just adds more debugging in case we ever see it again.

Note that a `data.Path` now needs to have a valid TLF ID set in order
to be considered valid.  This is a change, but I don't think it should
change any existing behavior.  Famous last words though...

Issue: HOTPOT-803